feat: Support Bitcoin Core v28
We add a feature flag `28_0` that is intended to provide
support for versions of Bitcoin Core from 28.0.

These methods are added to `client::v28` module

- get_block_header_verbose
- get_block_verbose

Also add feature flag `29_0`. This is currently the
default feature. It's recommended to specify at least 1
version feature that matches the version of Bitcoin Core
you're running. Features aren't necessarily additive.
In case multiple features are specified, the library must
respect the lowest version.